# [chore] [exporterhelper] Integrate capacity limiting into the communication channel