--- a/exporter/internal/queue/queue.go
+++ b/exporter/internal/queue/queue.go
@@ -34,3 +34,26 @@ type Queue[T any] interface {
 	// Capacity returns the capacity of the queue.
 	Capacity() int
 }
+
+type itemsCounter interface {
+	ItemsCount() int
+}
+
+// Sizer is an interface that returns the size of the given element.
+type Sizer[T any] interface {
+	Sizeof(T) int64
+}
+
+// ItemsSizer is a Sizer implementation that returns the size of a queue element as the number of items it contains.
+type ItemsSizer[T itemsCounter] struct{}
+
+func (is *ItemsSizer[T]) Sizeof(el T) int64 {
+	return int64(el.ItemsCount())
+}
+
+// RequestSizer is a Sizer implementation that returns the size of a queue element as one request.
+type RequestSizer[T any] struct{}
+
+func (rs *RequestSizer[T]) Sizeof(T) int64 {
+	return 1
+}