… exporter/internal/queue/queue_capacity.go
… exporter/internal/queue/queue_capacity_test.go
… exporter/internal/queue/sized_channel_test.go